requirements.txt
asgiref==3.4.1
cycler==0.10.0
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.0
Pillow==8.3.0
psycopg2==2.9.1
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.2.0
